==> scripts/script2/13_visualisation_interractive.R <==
# ==============================================================================
#            13_visualisation_interractive
# ==============================================================================

# Auteur: HAMID DIGORGORD
# Date: 2024-12-16

# -----------------------------------------------------------------------------
#                    1. CONFIGURATION INITIALE
# -----------------------------------------------------------------------------

# Chargement des packages
library(tidyverse)
library(here)

# -----------------------------------------------------------------------------
#                    2. FONCTIONS PRINCIPALES
# -----------------------------------------------------------------------------

# -----------------------------------------------------------------------------
#                    3. EXÉCUTION PRINCIPALE
# -----------------------------------------------------------------------------

main <- function() {
    message('Début de l'exécution...')
    # Ajouter le code ici
}

# Exécution directe
if (!interactive()) main()
